fix participants tests
